Fixed styles & added media
