+++ src/components/Sidebar/Sidebar.styled.ts
@@ -5,11 +5,15 @@ export const SidebarWrapper = styled.aside`
 	flex-direction: column;
 	gap: 1rem;
 
-	padding: 2rem 0 2rem 1rem;
-`;
+	padding: 2rem;
+	margin: 0 auto;
 
-export const LabelsContainer = styled.ul`
-	display: flex;
-	flex-direction: column;
-	gap: 0.5rem;
+	width: 100%;
+	max-width: 500px;
+
+	@media (min-width: 1100px) {
+		padding: 2rem 0 2rem 1rem;
+
+		max-width: auto;
+	}
 `;
