+++ src/components/UI/index.ts
@@ -1,3 +1,4 @@
 export * from './Button.styled';
 export * from './FieldLabel.styled';
 export * from './Text.styled';
+export * from './TextInput.styled';
